% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tilt_create_variable_labels}
\alias{tilt_create_variable_labels}
\title{Add a column of axis labels for current variables}
\usage{
tilt_create_variable_labels(dat, convert_to_ordered_factor = TRUE)
}
\arguments{
\item{dat}{Data frame of tilt meter data in long format.}

\item{convert_to_ordered_factor}{Logical variable indicating whether the new
\code{variable_label} column should be converted to an ordered factor.
Default is \code{TRUE}.}
}
\value{
Returns \code{dat} with an additional \code{variable_label} column
  for use in faceted figures.
}
\description{
Add a column of axis labels for current variables
}
