% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tilt_export_path}
\alias{tilt_export_path}
\title{Generate file path and name to export tilt meter data}
\usage{
tilt_export_path(depl_info, path = NULL, sub_folder = NULL, ext = "rds")
}
\arguments{
\item{depl_info}{Row of the tilt meter metadata tracking sheet that
corresponds to the deployment of interest. Must include columns
\code{county}, \code{station}, \code{deployment_date}, and
\code{deployment_id}.}

\item{path}{File path to a folder named \code{county}.}

\item{sub_folder}{Character string of the sub-folder name (inside county
folder) where \code{dat} should be exported. Default is \code{sub-folder =
"new"}.}

\item{ext}{File extension. Default is \code{ext = "csv"}.}
}
\value{
A file path for exporting deployment data, including file name and
  extension.
}
\description{
Generate file path and name to export tilt meter data
}
