% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilt_plot_flag_vars.R
\name{tilt_ggplot_flags}
\alias{tilt_ggplot_flags}
\title{Create ggplot for one wave qc_test and variable}
\usage{
tilt_ggplot_flags(
  dat,
  qc_test,
  point_size,
  n_col = NULL,
  flag_title = TRUE,
  plotly_friendly = FALSE
)
}
\arguments{
\item{dat}{Data frame of flagged wave data in long format. Must
include a column named with the string "_flag_value".}

\item{qc_test}{qc test to plot.}

\item{point_size}{Numeric argument indicating size of the points. Passed to
\code{geom_point()}.}

\item{n_col}{Number of columns for faceted plots.}

\item{flag_title}{Logical argument indicating whether to include a ggtitle of
the qc test and variable plotted.}

\item{plotly_friendly}{Logical argument. If \code{TRUE}, the legend will be
plotted when \code{plotly::ggplotly} is called on \code{p}. Default is
\code{FALSE}, which makes the legend look better in a static figure.}
}
\value{
Returns a ggplot object; a figure for \code{qc_test}.
  Points are coloured by the flag value and panels are faceted by depth and
  sensor.
}
\description{
Create ggplot for one wave qc_test and variable
}
