% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilt_pivot_longer.R
\name{tilt_pivot_longer}
\alias{tilt_pivot_longer}
\title{Pivot title meter data variables longer}
\usage{
tilt_pivot_longer(dat)
}
\arguments{
\item{dat}{Data frame of tilt meter data.}
}
\value{
Returns dat in a long format, with variable as an ordered factor.
}
\description{
Pivot title meter data variables longer
}
