% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilt_plot_ts.R
\name{tilt_plot_ts}
\alias{tilt_plot_ts}
\title{Plot tilt meter parameters over time}
\usage{
tilt_plot_ts(
  dat,
  geom = "line",
  pal = NULL,
  n_col = 1,
  size = 1,
  scales = "fixed",
  show_legend = FALSE
)
}
\arguments{
\item{dat}{Data frame with tilt meter data in wide or long format.}

\item{geom}{Character string indicating whether to plot line or points.
Options are "point" or "line". Default is "line".}

\item{pal}{Optional colour palette. Default is c("#104862", "#62AECA",
"#FFD118") (Indigo dye, Moonstone, Jonquil). For plots with more than 3
variables, these colours are interpolated using
\code{grDevices::colorRampPalette}.}

\item{n_col}{Number of columns for faceted figure. Default is 1.}

\item{size}{Numeric value specifying the line width or point size.}

\item{scales}{Character string indicating how to treat scales of the faceted
plot. Passed to \code{facet_wrap()}. Default is "fixed". Other options are
"free_x" and "free_y".}

\item{show_legend}{Logical argument indicating whether to show figure legend.
Default is \code{FALSE}.}
}
\value{
Returns a ggplot object of the plotted variables.
}
\description{
Plot tilt meter parameters over time
}
