% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilt_read_data.R
\name{tilt_read_current_data}
\alias{tilt_read_current_data}
\title{Read csv file of tilt meter current data}
\usage{
tilt_read_current_data(path, files, cmar_header = TRUE)
}
\arguments{
\item{path}{File path to the current and temperature csv files.}

\item{files}{Character vector with the names of the current and temperature
csv files, including the file extension.}

\item{cmar_header}{Logical argument indicating whether to replace original
file headers with standard CMAR column names.}
}
\value{
Returns data frame of tilt meter current data (speed and direction).
}
\description{
Read csv file of tilt meter current data
}
