% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilt_read_data.R
\name{tilt_read_data}
\alias{tilt_read_data}
\title{Read in current and temperature tilt meter data}
\usage{
tilt_read_data(path, files, timestamp_to_utc = TRUE, cmar_header = TRUE)
}
\arguments{
\item{path}{File path to the current and temperature csv files.}

\item{files}{Character vector with the names of the current and temperature
csv files, including the file extension.}

\item{timestamp_to_utc}{Logical argument indicating whether to convert the
  timestamp column from Halifax time to UTC. \code{cmar_header} must also be
  set to \code{TRUE} for this to work.

  For the raw tilt meter data, the timestamp column is in the timezone of the
  deployment date (e.g., "AST" if deployed in November to March and "DST" if
  deployed in March to November). The timestamp does NOT account for changes
  in daylight savings time.

  When \code{timestamp_to_utc = TRUE}, the timestamp is converted to UTC by
  adding 3 hours if the deployment date was during daylight savings, or 4
  hours if the deployment date was during Atlantic Standard Time.

  The earliest timestamp is used to define the original timezone (AST/DST).}

\item{cmar_header}{Logical argument indicating whether to replace original
file headers with standard CMAR column names.}
}
\value{
Returns a data frame with tilt meter current and temperature data,
  joined by timestamp.
}
\description{
Read in current and temperature tilt meter data
}
