% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilt_read_tracking.R
\name{tilt_read_tracking}
\alias{tilt_read_tracking}
\title{Read tilt meter deployment metadata}
\usage{
tilt_read_tracking(path = NULL, file = NULL, sheet = NULL)
}
\arguments{
\item{path}{File path to the folder with the metadata file. Default is the
metadata_tracking folder on the CMAR R drive.}

\item{file}{Name of the metadata file, includeing file extension. Must be a
".xlsx" file. Default is "tilt_meter_deployment_tracking.xlsx".}

\item{sheet}{Name of the sheet to read in.}
}
\value{
Returns a data frame of the metadata tracking sheet.
}
\description{
Read tilt meter deployment metadata
}
