% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilt_test_grossrange.R
\name{tilt_test_grossrange}
\alias{tilt_test_grossrange}
\title{Apply grossrange test to tilt meter variables}
\usage{
tilt_test_grossrange(
  dat,
  sensor_type = c("TCM-1", "TCM-4"),
  tilt_grossrange_table = NULL
)
}
\arguments{
\item{dat}{Data frame of tilt meter variables in wide format.}

\item{sensor_type}{Character string indicating which sensor collected the
data. Options are "TCM-1" or "TCM-4".}

\item{tilt_grossrange_table}{Data frame with at least 5 columns:
\code{variable}: entries must match the names of the variables being tested
in \code{dat}; \code{sensor_min}: minimum acceptable value;
\code{sensor_max}: maximum accepted value ; \code{user_min}: minimum
reasonable value; \code{user_max}: maximum reasonable value.}
}
\value{
Returns \code{dat} with an additional grossrange_flag column for each
  tilt meter variable.
}
\description{
Apply grossrange test to tilt meter variables
}
