% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilt_thresholds.R
\docType{data}
\name{tilt_thresholds}
\alias{tilt_thresholds}
\title{Threshold tables}
\format{
Default thresholds for adding flags to water quality data.
}
\source{
Add references here
}
\usage{
tilt_thresholds
}
\description{
Threshold tables
}
\keyword{datasets}
