% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilt_write_report_table.R
\name{tilt_write_report_table}
\alias{tilt_write_report_table}
\title{Writes deployment table for summary report}
\usage{
tilt_write_report_table(metadata)
}
\arguments{
\item{metadata}{Deployment metadata from the tilt meter metadata tracking
sheet.}
}
\value{
Returns a tibble with columns for the report table.
}
\description{
Writes deployment table for summary report
}
